Split parsing into tables module.
